1. Из MainController выделены UsersController и AuthoritiesController. 2. Исправлены ошибки в шаблонах. Теперь корректно работают стили и скрипты.
